Merge .pkgsave files based on etcupdate database
When pkg overwrites a file while installing a pkg it creates a .pkgsave
file (if the new file is not identical to the old).

To handle this robustly and automatically, pkgbasify now performs a
3-way-merge. The etcupdate database conveniently contains the necessary
shared ancestor of the .pkgsave file and the new file installed by pkg.

Sponsored by:	The FreeBSD Foundation