Do not rely on GC to close files
Relying on the GC to close files can cause hard to debug issues,
especially when combined with the buffering that is done by
libc's FILE that lua's file interface wraps.

The case that bit me was writing a configuration file to be read by a
subsequent pkg command but not closing the file before using
os.execute(). This was made more confusing due to the fact that adding
debugging code could cause the GC to be run soon enough and close the
file before the os.execute() call and obscuring the problem.

Sponsored by:	The FreeBSD Foundation