Fix latent bug in capture()
Currently capture() returns nil if the last character of the output is
not a newline. This is not the intended behavior and contradicts the doc
comments, but hasn't manifested as a user-facing bug yet since unixy
tools terminate their output with a new line by convention.

Sponsored by:	The FreeBSD Foundation