Add --force option to bypass already_pkgbase()
If the final pkg install fails due to e.g. lack of disk space, the
system may be left in a partially converted state where FreeBSD-runtime
has already been installed but not all other pkgbase packages have.

In this case, the user will likely want to run pkgbasify again to
install the rest of the packages, but currently this is impossible
without editing the script due to the already_pkgbase() check.

This commit adds a --force option to support this use case.

Sponsored by:	The FreeBSD Foundation
Closes:		https://github.com/FreeBSDFoundation/pkgbasify/issues/11